Add swift support for mixed objc-swift target
This will only add support for referencing swift from objc code, the other way will follow up in different commit
