+++ src/com/facebook/buck/cxx/BUCK.autodeps
@@ -1,4 +1,4 @@
-#@# GENERATED FILE: DO NOT MODIFY dbc88f6060f1a52a53cce3f9063da85364ef5438 #@#
+#@# GENERATED FILE: DO NOT MODIFY b5d7e5439e8c34a09f64ad08c093e728f8206895 #@#
 {
   "framework_path_arg" : {
     "deps" : [ ],
@@ -44,6 +44,7 @@
       "//src/com/facebook/buck/log:api",
       "//src/com/facebook/buck/rules:command_tool",
       "//src/com/facebook/buck/step/fs:fs",
+      "//src/com/facebook/buck/swift:swift",
       "//src/com/facebook/buck/test/result/type:type",
       "//src/com/facebook/buck/util:escaper",
       "//src/com/facebook/buck/util:io",
